+++ bench.py
@@ -10,19 +10,22 @@ METS = []
 for x in xrange(NUM):
     key = random.choice(KEYS)
     val = random.choice(VALS)
-    METS.append("%s:%f|c|@123\n" % (key, val))
+    METS.append("%s:%f|c\n" % (key, val))
 
 s = socket.socket()
 s.connect(("localhost", 8125))
 start = time.time()
 
-current = 0
-while current < len(METS):
-    msg = "".join(METS[current:current + 1024])
-    current += 1024
-    s.sendall(msg)
+total = 0
+while True:
+    current = 0
+    while current < len(METS):
+        msg = "".join(METS[current:current + 1024])
+        current += 1024
+        total += 1024
+        s.sendall(msg)
 
-s.close()
-end = time.time()
-print NUM / (end - start), "ops/sec", (end - start), "sec"
+    diff = time.time() - start
+    ops_s = total / diff
+    print "%0.2f sec\t - %.0f ops/sec" % (diff, ops_s)