+++ src/histogram.h
@@ -1,0 +1,8 @@
+
+typedef struct {
+    double min_val;
+    double max_val;
+    double bucket_width;
+    int num_buckets;
+} histogram_config;
+